backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.4.1
anthropic==0.49.0
mistralai==1.5.0
python-dotenv==1.0.1
requests==2.32.3
simple-websocket==1.1.0
